fix : turning single accordion item into an actual component + removing useless comments
