+++ src/MyAccordion.js
@@ -7,37 +7,14 @@ import myAcordionStyles from "./myAcordionStyles.module.css";
 import { BsPlayCircleFill } from "react-icons/bs";
 import { AiOutlineFile } from "react-icons/ai";
 import { createMuiTheme } from "@mui/material";
-
+import AccordionItem from "./AccordionItem";
 function MyAccordion(props) {
-  function createItem(item) {
-    let res = [];
-    let icon = item.is_coding_exercise ? (
-      <AiOutlineFile />
-    ) : (
-      <BsPlayCircleFill className={myAcordionStyles["play"]} />
-    );
+  let hours =
+    Math.round(props.data.content_length / 3600) > 0
+      ? Math.round(props.data.content_length / 3600) + "hrs  "
+      : "";
+  let mins = Math.round(props.data.content_length / 60) % 60;
 
-    let name = item.can_be_previewed ? (
-      <span className={myAcordionStyles["purple"]}> {item.title}</span>
-    ) : (
-      <span className={myAcordionStyles["black"]}> {item.title}</span>
-    );
-    let preview = item.can_be_previewed ? (
-      <span className={myAcordionStyles["preview"]}>Preview</span>
-    ) : (
-      ""
-    );
-    let time = (
-      <span className={myAcordionStyles["grey"]}>{item.content_summary}</span>
-    );
-    res.push(icon);
-    res.push(name);
-    res.push(preview);
-    res.push(time);
-    //console.log(res);
-    return <div className={myAcordionStyles["details-row"]}>{res}</div>;
-  }
-  //console.log(props);
   return (
     <Accordion className={myAcordionStyles["accordion"]}>
       <AccordionSummary
@@ -51,20 +28,14 @@ function MyAccordion(props) {
             {props.data.title}
           </h4>
           <span className={myAcordionStyles["duration"]}>
-            {props.data.lecture_count} Lectures •{" "}
-            {Math.round(props.data.content_length / 3600) > 0
-              ? Math.round(props.data.content_length / 3600) + "hrs  "
-              : ""}
-            {Math.round(props.data.content_length / 60) % 60}min
+            {props.data.lecture_count} Lectures • {hours}
+            {mins}min
           </span>
         </div>
       </AccordionSummary>
       <AccordionDetails className={myAcordionStyles["accordion-details"]}>
         {props.data.items.map((item) => {
-          //console.log(item);
-          let itemjsx = createItem(item);
-          //console.log(itemjsx);
-          return itemjsx;
+          return <AccordionItem data={item} />;
         })}
       </AccordionDetails>
     </Accordion>
